knowledge_agent/requirements.txt
requests==2.31.0
openai==1.5.0
chromadb==0.4.18
sentence-transformers==2.2.2
markdown==3.5.1
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
pydantic==2.5.0
neo4j==5.15.0
tiktoken==0.5.2
numpy==1.24.3
pandas==2.1.4
streamlit==1.29.0
plotly==5.17.0
networkx==3.2.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
websockets==12.0
watchdog==6.0.0